In, this commit, I to figure out how to place the image, describing my problem onto the README file
'Tis True! It took me a minute to figure out how to place the image I am requesting to be placed onto the README file onto the README file. This is quite a great breakthrough that I have discovered, and I'm very pleased with the future implications that this may bring.